requirements.txt
streamlit==1.32.0
torch==2.1.0
numpy==1.24.0
pandas==2.0.0
scipy==1.11.0
scikit-learn==1.3.0
PyMuPDF==1.23.0          # fitz - baca PDF
Pillow==10.0.0
matplotlib==3.7.0
plotly==5.17.0
